factor db.py code
